Get the monorepo to build
Remove examples and documentation; they were generating build errors, and it's not worth fixing.

Hard-code environment variables (using a Windows-incompatible format for now, for simplicity) to work around OpenSSL errors in newer Node.js versions and to silence Create React App errors about version mismatches. (Apparently this version of Lerna and/or this monorepo doesn't pin packages.)

Add skipLibCheck.

Pin specific versions of Vue and @types/babel__traverse to accommodate the old version of TypeScript that the monorepo is using.